Match patterns to input form user
The patterns available now are primative and only apply to standard
'hello' and 'bye' style messages.  When one of those two patterns is
recognized the app responds accordingly, including exiting the app
when the user says 'bye' or 'exit'.  A default case was also added
incase the input from the user is not matched to any pattern.